Add high-pass filter for NTC/Speed Hall multiplex
When Speed Sensor is Multiplexed with NTC 10k thermistor like
on motors with Z910 connector, we need to ignore low ADC values
when the hall crosses the magnet and speed sensor digital pin
is pulled down by hall sensor. With 220k pull-up resistor a good
starting point is to set SP_TEMP_MULTIPLEX to 150.

For this setup to work, connect the SS1 pin on controller to
MOTOR_TEMP pin.

If real NTC 10k resistance is close to 150, your motor is on fire